+++ packages/react-ui/storybook/stories/widgetsUI/FormulaWidgetUI.stories.js
@@ -1,5 +1,6 @@
 import React from 'react';
-import FormulaWidgetUI from '../../../src/widgets/FormulaWidgetUI';
+import FormulaWidgetUI from '../../../src/widgets/FormulaWidgetUI/FormulaWidgetUI';
+import { Label, ThinContainer } from '../../utils/storyStyles';
 
 const options = {
   title: 'Organisms/Widgets/FormulaWidgetUI',
@@ -22,7 +23,27 @@ const options = {
 
 export default options;
 
-const Template = (args) => <FormulaWidgetUI {...args} />;
+const Template = (args) => {
+  return <FormulaWidgetUI {...args} />;
+};
+
+const LoadingTemplate = (args) => {
+  return (
+    <>
+      <Label variant='body1' mb={3}>
+        {'Limited width'}
+      </Label>
+      <ThinContainer>
+        <FormulaWidgetUI {...args} />
+      </ThinContainer>
+
+      <Label variant='body1' mt={8} mb={3}>
+        {'Responsive'}
+      </Label>
+      <FormulaWidgetUI {...args} />
+    </>
+  );
+};
 
 const data = 10000;
 
@@ -54,3 +75,7 @@ const formatter = (value) =>
   }).format(value);
 const FormatterTextProps = { data, formatter };
 FormatterText.args = FormatterTextProps;
+
+export const Loading = LoadingTemplate.bind({});
+const LoadingProps = { data, isLoading: true };
+Loading.args = LoadingProps;